+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+## 2.0.0-rc5
+
   - totally reworked multiple fields (#542):
     - created a new Class: `ParsleyFieldMultiple`
     - multiple fields returns now same `ParsleyFieldMultiple` instance